+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Disabled `C-S-n` in `VSCode`
+- Use `C-S-k` to clear terminals in `VSCode`
+- Disabled `C-k` passthrough to `neovim` in `VSCode`
 - Added `vsliveshare` for `work`
 - Replaced `mongodb-tools` with `mongodb-compass` for `work`
 - Fixed `sigma` failing to boot by using an old version of `bcachefs`